Created publisher user controller and refacttiring code
